fix: green the CI lint gate (golangci-lint v7/v2.8.0, burn down 48 issues) (#2)
The CI 'Run linting' step had never passed. Two root causes, both fixed.

1. Workflow pinned golangci-lint-action@v6, which installs the v1 linter
   line. The repo's .golangci.yml is v2 and go.mod targets go1.25, so v1
   errored at config load. Bump to @v7 + version v2.8.0 (matches the local
   toolchain dev runs).

2. .golangci.yml declared version "2" but used v1 schema keys
   (linters-settings, issues.exclude-dirs), so v2 silently ignored the
   gosec excludes, errcheck/govet settings, and dir exclusions. Migrated to
   the v2 schema (linters.settings, linters.exclusions.paths). This made the
   maintainer's already-declared policy actually take effect (G304 excluded,
   check-type-assertions on). enable-all govet surfaced only fieldalignment
   + shadow — disabled both with justification (field order encodes logical
   grouping; err-shadowing is idiomatic Go).

Burned down all remaining issues by fixing the underlying cause, not
suppressing:
- audit.Log failures were swallowed at 10 call sites: added one logAudit
  seam that surfaces write failures to the operational log, and removed the
  dead `if auditLogger != nil` guards (the Noop()/disabled-logger pattern
  already makes nil impossible).
- certcache: replaced two untyped sync.Maps + unchecked type assertions with
  typed maps under one mutex; holding the lock across check-and-claim makes
  single-flight atomic and removes the dead double-check branch. Verified
  under -race (TestCertificateConcurrentAccess/Caching).
- recorder: surface the first JSONL write failure at Close so a truncated
  recording is never silent.
- shutdown/log-file close errors in run/examine now logged, not dropped.
- read-side Body.Close / best-effort symlink remove get explicit `_ =`.
- G115: index the palette with the uint64 directly (no narrowing convert).
- G101: targeted nolint on an event-name constant (false positive).
- staticcheck QF1012/QF1003/ST1005, ineffassign, unused: real fixes.

Acceptance: `make lint` exits 0; build, gofmt, and `make test-race` all pass.

+++ internal/log/logger.go
@@ -56,7 +56,9 @@ var colorIndex atomic.Uint64
 // NextCorrelationColor returns the next color from the palette for request/response correlation.
 // Rotates through the palette to ensure adjacent connections have different colors.
 func NextCorrelationColor() (r, g, b int) {
-	idx := int(colorIndex.Add(1)) % len(colorPalette)
+	// Reduce modulo the palette length in uint64 space and index directly —
+	// a uint64 is a valid slice index, so no narrowing conversion is needed.
+	idx := colorIndex.Add(1) % uint64(len(colorPalette))
 	c := colorPalette[idx]
 	return c[0], c[1], c[2]
 }
